processing/requirements.txt
apache-flink==1.20.0
google-cloud==0.34.0
google_cloud_bigquery_storage==2.26.0
protobuf
cloudpickle==2.2.1
apache-beam==2.48.0
python-dotenv==1.0.1